added time slider into the ui
