+++ client/components/time-field.js
@@ -1,15 +1,14 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
 import {connect} from 'react-redux'
 import {makeStyles} from '@material-ui/core/styles'
 import IconButton from '@material-ui/core/IconButton'
 import {PlayCircleOutline} from '@material-ui/icons'
-import {Slider} from '@material-ui/lab'
+import {TextField} from '@material-ui/core'
 import Typography from '@material-ui/core/Typography'
 
 const useStyles = makeStyles(theme => ({
   root: {
-    width: '60%',
+    width: '15%',
     padding: 5,
     backgroundColor: 'white',
     position: 'fixed',
@@ -27,58 +26,54 @@ const useStyles = makeStyles(theme => ({
 
 const marks = [
   {
-    value: 0,
+    value: '01:00',
     label: '12 AM - 2 AM'
   },
   {
-    value: 9.09,
+    value: '03:00',
     label: '2AM - 4AM'
   },
   {
-    value: 18.18,
+    value: '05:00',
     label: '4AM - 6AM'
   },
   {
-    value: 27.27,
+    value: '07:00',
     label: '6AM - 8AM'
   },
   {
-    value: 36.36,
+    value: '09:00',
     label: '8AM - 10AM'
   },
   {
-    value: 45.45,
+    value: '11:00',
     label: '10AM - 12PM'
   },
   {
-    value: 54.54,
+    value: '13:00',
     label: '12PM - 2PM'
   },
   {
-    value: 63.63,
+    value: '15:00',
     label: '2PM - 4PM'
   },
   {
-    value: 72.72,
+    value: '17:00',
     label: '4PM - 6PM'
   },
   {
-    value: 81.81,
+    value: '19:00',
     label: '6PM - 8PM'
   },
   {
-    value: 90.9,
+    value: '21:00',
     label: '8PM - 10PM'
   },
   {
-    value: 99.99,
+    value: '23:00',
     label: '10PM - 12AM'
   }
 ]
-
-function valueLabelFormat(value) {
-  return marks.findIndex(mark => mark.value === value) + 1
-}
 
 function DiscreteSlider(props) {
   const classes = useStyles()
@@ -106,20 +101,13 @@ function DiscreteSlider(props) {
         </IconButton>
       </Typography>
       <div style={{display: 'flex'}}>
-        <Slider
-          className={classes.slider}
-          value={hourPct}
-          valueLabelFormat={valueLabelFormat}
-          aria-labelledby="discrete-slider-restrict"
-          step={null}
-          marks={marks}
-          onChangeCommitted={(e, v) => {
-            if (
-              Math.round(24 * (v / 100)) !==
-              Math.round(24 * (props.currentHourPct / 100))
-            )
-              return props.changeTime(Math.round(24 * (v / 100)))
+        <TextField
+          value={marks}
+          onChange={(e, v) => {
+            console.log(e, v)
+            props.changeTime(v)
           }}
+          type="time"
         />
       </div>
     </div>
