+++ public/style.css
@@ -1,3 +1,7 @@
 body {
   margin: 0px;
 }
+
+.makeStyles-root-1 {
+  width: 1000px;
+}
